+++ HashtagHelp/Domain/Models/HashtagEntity.cs
@@ -4,6 +4,6 @@
     {
         public string Name { get; set; } = string.Empty;
         public Guid Id { get; set; } = Guid.NewGuid();
-        public uint MediaCount { get; set; }
+        public long MediaCount { get; set; }
     }
 }
